Complementary files
These two files are needed for the implementation of the linear VProfile. Replace them into the right folder before compiling the .cpp file (see last commit). 

These files contain the definition of the functor called LinearVProfile. It is composed of a class (which is in the .h file) and an operator (which is in the .hh file). The class contains the constructor --> the one used to call the functor. The operator contains the analytical definition of the VProfile as such.